Migrate disagg+SpD tests onto tests/e2e/ infra
Moves the disagg+SpD (draft_model/ngram/suffix) hardware E2E suite from
the old hand-rolled CLI-option test tree onto the tests/e2e/ marker/fixture
pattern established in 4a18ee6, resolving a latent duplicate-pytest_addoption
conflict between the two trees. Pure-Python guard tests are split out into
their own file since they need none of the device-pool machinery.
